+++ fe/src/pages/StoreDetail/StoreDetail.tsx
@@ -1,8 +1,8 @@
 import { DEFAULT_IMG, FEEDS, STORE } from 'constants/dummyData';
 import { ICON, STORE_DETAIL } from 'constants/size';
 import TEXT from 'constants/text';
-import Icon from 'common/Icon';
-import CommonHeader from 'components/CommonHeader/CommonHeader';
+import CommonHeader from 'components/Common/Header/CommonHeader';
+import Icon from 'components/Common/Icon/Icon';
 import Feed from 'components/ReviewList/Feed/Feed';
 import {
   DimText,
